requirements.txt
toml
numpy
uvicorn
fastapi
Pillow
requests
pyyaml
python-multipart
tinydb
ordered-set
dill
peewee
celery
redis
qrcode
BeautifulSoup4
lxml
oss2
alibabacloud_imageseg20191230==3.0.0
cachetools
pymongo
injector
requests_toolbelt